backend/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
google-api-python-client==2.126.0
pypdf==4.2.0